rename the folder again
