+++ src/App.tsx
@@ -2,7 +2,7 @@ import "bulma/css/bulma.css";
 import "@fortawesome/fontawesome-free/css/all.css";
 import "./App.scss";
 import { Outlet } from "react-router-dom";
-import NavBar from "./comp/NavBar";
+import NavBar from "./component/NavBar";
 
 export const App = () => (
   <>
